Implement scrolling shelf for AnimeShelf, ChipShelf
There are still a couple desktop UX issues (hovering over cards makes them chip on the scroll container, and I can't quite get the 'scroll' buttons to look as good yet), so I will finalize implementation for desktop in a follow-up.
