+++ package.json
@@ -23,6 +23,7 @@
     "react-dom": "^18.2.0",
     "react-firebase-hooks": "^5.1.1",
     "react-hook-form": "^7.42.1",
+    "react-horizontal-scrolling-menu": "^4.0.4",
     "react-router-dom": "^6.4.5",
     "react-scripts": "5.0.1",
     "react-swipeable": "^7.0.0",
