+++ src/Components/ShelfTitle.js
@@ -17,7 +17,7 @@ export default function ShelfTitle({ selectedGenre, setSelectedGenre, title }) {
       }}
     >
       <Container maxWidth="lg">
-        <Typography variant="h3" sx={{ marginBottom: "0.3em" }}>
+        <Typography variant="h3" sx={{ marginBottom: "0.125em" }}>
           {title}
         </Typography>
 
@@ -26,8 +26,6 @@ export default function ShelfTitle({ selectedGenre, setSelectedGenre, title }) {
           setSelectedGenre={setSelectedGenre}
         />
       </Container>
-
-      <div style={{ marginTop: "-15px", height: "1px" }}></div>
     </div>
   );
 }
